% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{API_MAX_BANKS_PER_REQUEST}
\alias{API_MAX_BANKS_PER_REQUEST}
\title{Maximum number of banks to fetch in single API call}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
API_MAX_BANKS_PER_REQUEST
}
\description{
Some API endpoints have limits on batch sizes.
Default: 30 banks
}
\keyword{internal}
