% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{API_RATE_LIMIT_DELAY}
\alias{API_RATE_LIMIT_DELAY}
\title{API rate limit delay in seconds}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
API_RATE_LIMIT_DELAY
}
\description{
Delay between API requests to avoid rate limiting.
Default: 0.05 seconds (50ms)
}
\keyword{internal}
