% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribits.R
\name{as_tibble.ribits_data}
\alias{as_tibble.ribits_data}
\title{Convert ribits_data to a tibble}
\usage{
as_tibble.ribits_data(x, ...)
}
\arguments{
\item{x}{A ribits_data object}

\item{...}{Additional arguments ignored}
}
\value{
A tibble
}
\description{
Returns the main \code{banks} (or \code{ilf}/\code{umbrellas}) dataframe.
This allows \code{ribits_data} objects to be used directly in tidyverse pipelines
that expect a dataframe.
}
\examples{
\dontrun{
ribits(state = "CA") |> as_tibble()
}
}
