% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{BANK_STATUS_EPA_LAYERS}
\alias{BANK_STATUS_EPA_LAYERS}
\title{EPA layer mapping for bank statuses}
\format{
An object of class \code{list} of length 6.
}
\usage{
BANK_STATUS_EPA_LAYERS
}
\description{
Maps each bank status to its corresponding EPA ArcGIS layer.
Note: The "terminated_banks" layer contains multiple statuses.
}
\keyword{internal}
