% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants}
\alias{constants}
\title{Package Constants}
\description{
This file defines all constants used throughout the RIBITSr package.
Centralizing constants makes them easier to find, document, and modify.
}
\keyword{internal}
