% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CSV_DOWNLOAD_DELAY}
\alias{CSV_DOWNLOAD_DELAY}
\title{CSV download delay in seconds}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CSV_DOWNLOAD_DELAY
}
\description{
Delay between CSV download requests.
Default: 0.5 seconds (500ms)
}
\keyword{internal}
