% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_API_TIMEOUT}
\alias{DEFAULT_API_TIMEOUT}
\title{Default API timeout in seconds}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_API_TIMEOUT
}
\description{
How long to wait for API responses before timing out.
Default: 30 seconds
}
\keyword{internal}
