% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_CACHE_AGE_DAYS}
\alias{DEFAULT_CACHE_AGE_DAYS}
\title{Default cache age in days}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_CACHE_AGE_DAYS
}
\description{
How long cached data remains valid before refresh.
Default: 30 days
}
\keyword{internal}
