% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_CHUNK_SIZE}
\alias{DEFAULT_CHUNK_SIZE}
\title{Default chunk size for batch processing}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_CHUNK_SIZE
}
\description{
Number of items to process in each batch when making bulk requests.
Default: 10
}
\keyword{internal}
