% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_CRS}
\alias{DEFAULT_CRS}
\title{Default CRS for spatial data}
\format{
An object of class \code{character} of length 1.
}
\usage{
DEFAULT_CRS
}
\description{
EPSG:4326 is WGS84 (GPS coordinates, lat/lon)
}
\keyword{internal}
