% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_CSV_TIMEOUT}
\alias{DEFAULT_CSV_TIMEOUT}
\title{Default CSV download timeout in seconds}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_CSV_TIMEOUT
}
\description{
CSV files can be large, so allow longer timeout.
Default: 300 seconds (5 minutes)
}
\keyword{internal}
