% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_MAX_CONCURRENT}
\alias{DEFAULT_MAX_CONCURRENT}
\title{Maximum concurrent requests}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_MAX_CONCURRENT
}
\description{
Maximum number of concurrent API requests when using parallel fetching.
Default: 5
}
\keyword{internal}
