% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_MAX_RETRIES}
\alias{DEFAULT_MAX_RETRIES}
\title{Default maximum retries for network requests}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_MAX_RETRIES
}
\description{
Number of times to retry failed requests before giving up.
Default: 5
}
\keyword{internal}
