% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_PREVIEW_ROWS}
\alias{DEFAULT_PREVIEW_ROWS}
\title{Default number of preview rows}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_PREVIEW_ROWS
}
\description{
Number of rows to show in data previews and samples.
Default: 5
}
\keyword{internal}
