% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_RETRY_DELAY}
\alias{DEFAULT_RETRY_DELAY}
\title{Default retry delay in seconds}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_RETRY_DELAY
}
\description{
Initial delay before first retry (doubles with each retry).
Default: 2 seconds
}
\keyword{internal}
