% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_SOURCE_PRIORITY}
\alias{DEFAULT_SOURCE_PRIORITY}
\title{Default priority order for data sources}
\format{
An object of class \code{character} of length 3.
}
\usage{
DEFAULT_SOURCE_PRIORITY
}
\description{
When discrepancies exist, this order determines which source to trust.
Default: CSV > API > EPA
Rationale:
\itemize{
\item CSV reports are official quarterly reports
\item API is real-time but sometimes has data quality issues
\item EPA ArcGIS can be outdated
}
}
\keyword{internal}
