% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_SPATIAL_FORMAT}
\alias{DEFAULT_SPATIAL_FORMAT}
\title{Default spatial export format}
\format{
An object of class \code{character} of length 1.
}
\usage{
DEFAULT_SPATIAL_FORMAT
}
\description{
Default spatial export format
}
\keyword{internal}
