% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_VERBOSE}
\alias{DEFAULT_VERBOSE}
\title{Default verbosity level}
\format{
An object of class \code{logical} of length 1.
}
\usage{
DEFAULT_VERBOSE
}
\description{
Controls how much progress information is displayed.
Can be overridden with rb_config(verbose = TRUE/FALSE)
}
\keyword{internal}
