% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribits.R
\name{discrepancies}
\alias{discrepancies}
\alias{discrepancies.ribits_data}
\title{Access discrepancies from ribits_data}
\usage{
discrepancies(x, ...)

\method{discrepancies}{ribits_data}(x, ...)
}
\arguments{
\item{x}{A ribits_data object}

\item{...}{Additional arguments passed to methods}
}
\value{
A tibble of discrepancies
}
\description{
Access discrepancies from ribits_data
}
