% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-columns.R
\name{.col_set}
\alias{.col_set}
\title{Safely Set Column in Dataframe}
\usage{
.col_set(df, col, value, case_insensitive = TRUE)
}
\arguments{
\item{df}{Dataframe to modify}

\item{col}{Column name to set}

\item{value}{Value to set}

\item{case_insensitive}{If TRUE, match existing columns case-insensitively}
}
\value{
Modified dataframe
}
\description{
Sets a column value in a dataframe, optionally using case-insensitive
matching to find existing columns.
}
\details{
If case_insensitive is TRUE and a column with different case exists,
that column will be updated. Otherwise, a new column is created.
}
\examples{
\dontrun{
df <- tibble::tibble(BankID = 1:3)

# Update existing column (case-insensitive)
df <- .col_set(df, "bank_id", c(10, 20, 30))  # Updates BankID

# Add new column
df <- .col_set(df, "status", c("active", "active", "inactive"))
}

}
\keyword{internal}
