% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization-compare.R
\name{.compare_values}
\alias{.compare_values}
\title{Compare two values and detect discrepancies}
\usage{
.compare_values(value1, value2, field_name, source1, source2, config)
}
\arguments{
\item{value1}{First value}

\item{value2}{Second value}

\item{field_name}{Name of the field being compared}

\item{source1}{Name of first source (e.g., "api")}

\item{source2}{Name of second source (e.g., "csv")}

\item{config}{Configuration from .get_discrepancy_config()}
}
\value{
NULL if values match, or a tibble with discrepancy details
}
\description{
Compare two values and detect discrepancies
}
\keyword{internal}
