% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization-report.R
\name{.create_discrepancy_summary}
\alias{.create_discrepancy_summary}
\title{Create summary sheet for Excel export}
\usage{
.create_discrepancy_summary(disc, resolutions)
}
\description{
Create summary sheet for Excel export
}
\keyword{internal}
