% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization-report.R
\name{.display_source_comparison}
\alias{.display_source_comparison}
\title{Display source comparison in CLI}
\usage{
.display_source_comparison(comparison, bank_id = NULL, field = NULL)
}
\description{
Display source comparison in CLI
}
\keyword{internal}
