% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-batch.R
\name{.estimate_batch_time}
\alias{.estimate_batch_time}
\title{Estimate Time for Batch Operation}
\usage{
.estimate_batch_time(
  n_items,
  batch_size = DEFAULT_CHUNK_SIZE,
  avg_time_per_batch = 0.5
)
}
\arguments{
\item{n_items}{Number of items to process}

\item{batch_size}{Size of each batch}

\item{avg_time_per_batch}{Average time per batch in seconds}
}
\value{
Estimated time in seconds
}
\description{
Estimates how long a batch operation will take based on batch size
and average request time.
}
\keyword{internal}
