% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization-report.R
\name{.export_discrepancies_csv}
\alias{.export_discrepancies_csv}
\title{Export discrepancies to CSV}
\usage{
.export_discrepancies_csv(disc, resolutions, file_path)
}
\description{
Export discrepancies to CSV
}
\keyword{internal}
