% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization-report.R
\name{.export_discrepancies_html}
\alias{.export_discrepancies_html}
\title{Export discrepancies to HTML table}
\usage{
.export_discrepancies_html(disc, resolutions, file_path)
}
\description{
Export discrepancies to HTML table
}
\keyword{internal}
