% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-messages.R
\name{.format_number}
\alias{.format_number}
\title{Format Number with Commas}
\usage{
.format_number(x)
}
\arguments{
\item{x}{Numeric value}
}
\value{
Character string with formatted number
}
\description{
Helper to format large numbers with thousand separators.
}
\keyword{internal}
