% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization-config.R
\name{.get_discrepancy_config}
\alias{.get_discrepancy_config}
\title{Get current discrepancy configuration}
\usage{
.get_discrepancy_config()
}
\value{
List with current configuration settings
}
\description{
Get current discrepancy configuration
}
\keyword{internal}
