% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization-resolve.R
\name{.get_source_values}
\alias{.get_source_values}
\title{Get source values for a specific bank and field}
\usage{
.get_source_values(matching_disc, matching_res, bank_id, field)
}
\description{
Get source values for a specific bank and field
}
\keyword{internal}
