% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize-transactions.R
\name{.harmonize_transactions}
\alias{.harmonize_transactions}
\title{Harmonize transaction data from multiple sources}
\usage{
.harmonize_transactions(api_ledger, csv_ledger)
}
\description{
Harmonize transaction data from multiple sources
}
\keyword{internal}
