% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization-resolve.R
\name{.load_manual_resolution_preferences}
\alias{.load_manual_resolution_preferences}
\title{Load saved manual resolution preferences}
\usage{
.load_manual_resolution_preferences()
}
\description{
Load saved manual resolution preferences
}
\keyword{internal}
