% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization-merge.R
\name{.merge_multiple_sources}
\alias{.merge_multiple_sources}
\title{Merge multiple data frames preserving all columns}
\usage{
.merge_multiple_sources(df_list, by = "bank_id", priority_order = NULL)
}
\arguments{
\item{df_list}{Named list of data frames to merge}

\item{by}{Column(s) to join on}

\item{priority_order}{Vector of names specifying merge priority}
}
\value{
A merged data frame
}
\description{
Merge multiple data frames preserving all columns
}
\keyword{internal}
