% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-messages.R
\name{.msg_progress_done}
\alias{.msg_progress_done}
\title{Complete Progress Bar}
\usage{
.msg_progress_done(id = NULL, quietly = NULL)
}
\arguments{
\item{id}{Progress bar ID (from .msg_progress_bar)}

\item{quietly}{If TRUE, suppress completion}
}
\description{
Complete Progress Bar
}
\keyword{internal}
