% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-messages.R
\name{.msg_progress_update}
\alias{.msg_progress_update}
\title{Update Progress Bar}
\usage{
.msg_progress_update(id = NULL, ..., quietly = NULL)
}
\arguments{
\item{id}{Progress bar ID (from .msg_progress_bar)}

\item{...}{Arguments passed to cli::cli_progress_update}

\item{quietly}{If TRUE, suppress update}
}
\description{
Update Progress Bar
}
\keyword{internal}
