% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-messages.R
\name{.msg_rate_limit}
\alias{.msg_rate_limit}
\title{API Rate Limit Message}
\usage{
.msg_rate_limit(delay, quietly = NULL)
}
\arguments{
\item{delay}{Delay in seconds}

\item{quietly}{If TRUE, suppress message}
}
\description{
Message for API rate limiting delays.
}
\keyword{internal}
