% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-messages.R
\name{.msg_validation_summary}
\alias{.msg_validation_summary}
\title{Validation Summary Message}
\usage{
.msg_validation_summary(n_passed, n_warnings, n_errors, quietly = NULL)
}
\arguments{
\item{n_passed}{Number of checks passed}

\item{n_warnings}{Number of warnings}

\item{n_errors}{Number of errors}

\item{quietly}{If TRUE, suppress message}
}
\description{
Display a summary of validation results.
}
\keyword{internal}
