% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-core.R
\name{.rb_get_next_link}
\alias{.rb_get_next_link}
\title{Get next link URL}
\usage{
.rb_get_next_link(data)
}
\arguments{
\item{data}{Parsed JSON response}
}
\description{
Get next link URL
}
\keyword{internal}
