% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-core.R
\name{.rb_has_next_link}
\alias{.rb_has_next_link}
\title{Check if response has next link}
\usage{
.rb_has_next_link(data)
}
\arguments{
\item{data}{Parsed JSON response}
}
\description{
Check if response has next link
}
\keyword{internal}
