% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{.rb_parse_geojson}
\alias{.rb_parse_geojson}
\title{Parse GeoJSON string to sf object}
\usage{
.rb_parse_geojson(geojson_string)
}
\arguments{
\item{geojson_string}{Character string containing GeoJSON.}
}
\value{
An sfc object or NULL if invalid.
}
\description{
Parse GeoJSON string to sf object
}
\keyword{internal}
