% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-messages.R
\name{.should_be_quiet}
\alias{.should_be_quiet}
\title{Get Current Verbosity Setting}
\usage{
.should_be_quiet(quietly = NULL)
}
\arguments{
\item{quietly}{Override for the current call}
}
\value{
Logical - TRUE if messages should be suppressed
}
\description{
Checks if messages should be displayed based on package configuration.
}
\keyword{internal}
