% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribits-methods.R
\name{filter.ribits_data}
\alias{filter.ribits_data}
\title{Filter ribits_data objects}
\usage{
\method{filter}{ribits_data}(.data, ...)
}
\arguments{
\item{.data}{A ribits_data object}

\item{...}{Filter expressions passed to dplyr::filter()}
}
\value{
A filtered ribits_data object
}
\description{
Apply dplyr::filter() to the banks dataframe within a ribits_data object.
This allows you to filter banks without accessing \verb{$banks} directly.
}
\examples{
\dontrun{
ca_banks <- ribits(state = "CA")
large_banks <- ca_banks |> filter(total_acres > 100)
}
}
