% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{get_error_type}
\alias{get_error_type}
\title{Get Error Type}
\usage{
get_error_type(error)
}
\arguments{
\item{error}{Error object}
}
\value{
Character string of error type, or NULL if not a RIBITSr error
}
\description{
Returns the specific type of RIBITSr error.
}
\keyword{internal}
