% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{HTTP_NO_RETRY_CODES}
\alias{HTTP_NO_RETRY_CODES}
\title{HTTP status codes that should not be retried}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
HTTP_NO_RETRY_CODES
}
\description{
These represent permanent failures that won't be fixed by retrying.
}
\keyword{internal}
