% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{HTTP_RETRY_CODES}
\alias{HTTP_RETRY_CODES}
\title{HTTP status codes that indicate temporary failures}
\format{
An object of class \code{numeric} of length 6.
}
\usage{
HTTP_RETRY_CODES
}
\description{
These may succeed if retried after a delay.
}
\keyword{internal}
