% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{HTTP_SUCCESS}
\alias{HTTP_SUCCESS}
\title{HTTP status codes for different response types}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
HTTP_SUCCESS
}
\description{
Used for classifying API responses and determining retry behavior.
}
\keyword{internal}
