% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MAX_CREDIT_VALUE}
\alias{MAX_CREDIT_VALUE}
\title{Maximum reasonable credit value}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MAX_CREDIT_VALUE
}
\description{
Credits above this are flagged for review (unusually large).
Default: 50000 acres
}
\keyword{internal}
