% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MAX_ESTABLISHMENT_YEAR_OFFSET}
\alias{MAX_ESTABLISHMENT_YEAR_OFFSET}
\title{Maximum reasonable establishment year}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MAX_ESTABLISHMENT_YEAR_OFFSET
}
\description{
Banks established after current year + this offset flagged as erroneous.
Default: 2 (allow 2 years in future for planning)
}
\keyword{internal}
