% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MAX_RETRY_DELAY}
\alias{MAX_RETRY_DELAY}
\title{Maximum retry delay in seconds}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MAX_RETRY_DELAY
}
\description{
Cap on exponential backoff to prevent excessive waits.
Default: 60 seconds
}
\keyword{internal}
