% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MAX_STRING_DISTANCE}
\alias{MAX_STRING_DISTANCE}
\title{Maximum string distance for fuzzy matching}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MAX_STRING_DISTANCE
}
\description{
Used in name matching when exact matches aren't found.
Lower values require closer matches.
Default: 0.2 (20\% difference allowed)
}
\keyword{internal}
