% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MIN_CREDIT_VALUE}
\alias{MIN_CREDIT_VALUE}
\title{Minimum reasonable credit value}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MIN_CREDIT_VALUE
}
\description{
Credits below this are flagged as potentially erroneous.
Default: 0 (no negative credits allowed)
}
\keyword{internal}
