% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MIN_CSV_FILE_SIZE}
\alias{MIN_CSV_FILE_SIZE}
\title{Minimum CSV file size in bytes}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MIN_CSV_FILE_SIZE
}
\description{
Files smaller than this are likely errors (empty or error pages).
Default: 100 bytes
}
\keyword{internal}
