% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MIN_ESTABLISHMENT_YEAR}
\alias{MIN_ESTABLISHMENT_YEAR}
\title{Minimum reasonable establishment year}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MIN_ESTABLISHMENT_YEAR
}
\description{
Banks established before this year are flagged as potentially erroneous.
Default: 1990 (RIBITS program started in early 1990s)
}
\keyword{internal}
