% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribits.R
\name{mutate.ribits_data}
\alias{mutate.ribits_data}
\title{Mutate ribits_data objects}
\usage{
\method{mutate}{ribits_data}(.data, ...)
}
\arguments{
\item{.data}{A ribits_data object}

\item{...}{Mutate expressions passed to dplyr::mutate()}
}
\value{
A ribits_data object with mutated columns
}
\description{
Apply dplyr::mutate() to the banks dataframe within a ribits_data object.
}
\examples{
\dontrun{
ca_banks <- ribits(state = "CA")
with_hectares <- ca_banks |> mutate(total_hectares = total_acres * 0.404686)
}
}
