% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{NAME_LOOKUP_CACHE_AGE_DAYS}
\alias{NAME_LOOKUP_CACHE_AGE_DAYS}
\title{Name lookup cache age in days}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
NAME_LOOKUP_CACHE_AGE_DAYS
}
\description{
How long name lookup tables are cached.
Default: 30 days (bank names change infrequently)
}
\keyword{internal}
