% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{NUM_HARMONIZATION_RULES}
\alias{NUM_HARMONIZATION_RULES}
\title{Number of auto-harmonization rules}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
NUM_HARMONIZATION_RULES
}
\description{
The package implements this many intelligent rules for resolving
data discrepancies automatically.
}
\keyword{internal}
