% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize-transactions.R
\name{print.ribits_transactions}
\alias{print.ribits_transactions}
\title{Print method for ribits_transactions}
\usage{
\method{print}{ribits_transactions}(x, ...)
}
\description{
Print method for ribits_transactions
}
