% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PROGRESS_BAR_MIN_DURATION}
\alias{PROGRESS_BAR_MIN_DURATION}
\title{Show progress bars for operations longer than this many seconds}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
PROGRESS_BAR_MIN_DURATION
}
\description{
Operations faster than this skip progress bars to reduce clutter.
Default: 2 seconds
}
\keyword{internal}
