% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-location.R
\name{rb_banks_by_location}
\alias{rb_banks_by_location}
\title{Find banks by location}
\usage{
rb_banks_by_location(lat, lon)
}
\arguments{
\item{lat}{Latitude (decimal degrees)}

\item{lon}{Longitude (decimal degrees)}
}
\value{
A tibble of banks near the location
}
\description{
Queries the RIBITS API to find mitigation banks near a given latitude/longitude.
}
\keyword{internal}
