% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-core.R
\name{rb_bool_to_text}
\alias{rb_bool_to_text}
\title{Convert logical to Yes/No for API parameters}
\usage{
rb_bool_to_text(x)
}
\arguments{
\item{x}{Logical value}
}
\value{
Character "Yes" or "No"
}
\description{
Convert logical to Yes/No for API parameters
}
\keyword{internal}
