% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{rb_checkpoints}
\alias{rb_checkpoints}
\title{List available checkpoints}
\usage{
rb_checkpoints()
}
\value{
A tibble of available checkpoints
}
\description{
Shows checkpoints from interrupted operations that can be resumed.
}
\examples{
\dontrun{
# See what can be resumed
rb_checkpoints()
}
}
