% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization-resolve.R
\name{rb_clear_manual_resolutions}
\alias{rb_clear_manual_resolutions}
\title{Clear all saved manual resolution preferences}
\usage{
rb_clear_manual_resolutions()
}
\value{
Invisibly returns TRUE if successful
}
\description{
Clear all saved manual resolution preferences
}
\examples{
\dontrun{
# Clear all saved override preferences
rb_clear_manual_resolutions()
}
}
