% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv-registry.R
\name{rb_csv_info}
\alias{rb_csv_info}
\title{Get CSV report info}
\usage{
rb_csv_info(report_type)
}
\arguments{
\item{report_type}{Character. The report type.}
}
\value{
A list with report metadata.
}
\description{
Returns metadata about a specific CSV report type.
}
\examples{
rb_csv_info("banks_sites")
rb_csv_info("credit_classification")
}
