% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrepancy-handling.R
\name{rb_export_discrepancies}
\alias{rb_export_discrepancies}
\title{Export discrepancies to CSV for manual review}
\usage{
rb_export_discrepancies(data, file_path = "ribits_discrepancies.csv")
}
\arguments{
\item{data}{A ribits_data object}

\item{file_path}{Path to save CSV file}
}
\description{
Export discrepancies to CSV for manual review
}
\keyword{internal}
