% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{rb_extract_contacts}
\alias{rb_extract_contacts}
\title{Extract contacts from bank object}
\usage{
rb_extract_contacts(bank)
}
\arguments{
\item{bank}{A tibble returned by \code{rb_get_bank()}}
}
\value{
A tibble containing contacts.
}
\description{
Extract contacts from bank object
}
