% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{rb_extract_footprint}
\alias{rb_extract_footprint}
\title{Extract footprint from RIBITS data}
\usage{
rb_extract_footprint(data)
}
\arguments{
\item{data}{A tibble returned by \verb{rb_get_*()} functions}
}
\value{
An sf object or empty tibble
}
\description{
Extract footprint from RIBITS data
}
\keyword{internal}
