% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{rb_extract_irt_members}
\alias{rb_extract_irt_members}
\title{Extract IRT members from RIBITS data}
\usage{
rb_extract_irt_members(data)
}
\arguments{
\item{data}{A tibble returned by \verb{rb_get_*()} functions}
}
\value{
A tibble
}
\description{
Extract IRT members from RIBITS data
}
\keyword{internal}
