% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{rb_extract_ledger}
\alias{rb_extract_ledger}
\title{Extract ledger from RIBITS data}
\usage{
rb_extract_ledger(data)
}
\arguments{
\item{data}{A tibble returned by \verb{rb_get_*()} functions}
}
\value{
A tibble containing the ledger
}
\description{
Convenience wrapper for \code{rb_extract(data, "ledger")}.
}
\keyword{internal}
