% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{rb_extract_managers}
\alias{rb_extract_managers}
\title{Extract managers from RIBITS data}
\usage{
rb_extract_managers(data)
}
\arguments{
\item{data}{A tibble returned by \verb{rb_get_*()} functions}
}
\value{
A tibble
}
\description{
Extract managers from RIBITS data
}
\keyword{internal}
