% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{rb_extract_other_contacts}
\alias{rb_extract_other_contacts}
\title{Extract other contacts from bank object}
\usage{
rb_extract_other_contacts(bank)
}
\arguments{
\item{bank}{A tibble returned by \code{rb_get_bank()}}
}
\value{
A tibble containing other contacts.
}
\description{
Extract other contacts from bank object
}
