% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{rb_extract_pocs}
\alias{rb_extract_pocs}
\title{Extract POCs from bank object}
\usage{
rb_extract_pocs(bank)
}
\arguments{
\item{bank}{A tibble returned by \code{rb_get_bank()}}
}
\value{
A tibble containing POCs.
}
\description{
Extract POCs from bank object
}
