% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilf_programs.R
\name{rb_extract_program_sites}
\alias{rb_extract_program_sites}
\title{Extract program sites from ILF program object}
\usage{
rb_extract_program_sites(ilf_program)
}
\arguments{
\item{ilf_program}{A tibble returned by \code{rb_get_ilf_program()}}
}
\value{
A tibble containing program sites.
}
\description{
Extract program sites from ILF program object
}
