% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rb_extract_service_area}
\alias{rb_extract_service_area}
\title{Extract service area}
\usage{
rb_extract_service_area(bank)
}
\arguments{
\item{bank}{A tibble returned by \code{rb_get_bank()}.}
}
\value{
An sf object containing the service area, or NULL if not found.
}
\description{
Extracts the service area geometry from a bank object and converts it to
an sf object.
}
