% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umbrellas.R
\name{rb_extract_umbrella_sites}
\alias{rb_extract_umbrella_sites}
\title{Extract umbrella sites from umbrella instrument object}
\usage{
rb_extract_umbrella_sites(umbrella)
}
\arguments{
\item{umbrella}{A tibble returned by \code{rb_get_umbrella()}}
}
\value{
A tibble containing umbrella sites.
}
\description{
Extract umbrella sites from umbrella instrument object
}
