% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banks.R
\name{rb_get_bank}
\alias{rb_get_bank}
\title{Get bank details}
\usage{
rb_get_bank(
  bank_id,
  service_area = TRUE,
  footprint = TRUE,
  contacts = TRUE,
  ledger = TRUE
)
}
\arguments{
\item{bank_id}{Integer. The unique RIBITS bank identifier.}

\item{service_area}{Logical. Include service area geometries? Default TRUE.}

\item{footprint}{Logical. Include bank footprint geometry? Default TRUE.}

\item{contacts}{Logical. Include contact information? Default TRUE.}

\item{ledger}{Logical. Include transaction ledger? Default TRUE.}
}
\value{
A list containing bank attributes and requested nested data.
}
\description{
Retrieves detailed information for a specific mitigation bank.
}
