% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribits-user.R
\name{rb_get_ledger}
\alias{rb_get_ledger}
\title{Get Master Ledger Dataframe}
\usage{
rb_get_ledger(aoi = NULL, ...)
}
\arguments{
\item{aoi}{Area of Interest (State, District, or Bank ID)}

\item{...}{Additional arguments passed to \code{ribits()}}
}
\value{
A tibble of transaction history
}
\description{
Convenience function to retrieve the "Master Ledger" dataframe.
Includes comprehensive transaction history from all sources (Watershed CSV + API + Ledger CSV).
}
