% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribits-user.R
\name{rb_get_summary}
\alias{rb_get_summary}
\title{Get Master Summary Dataframe}
\usage{
rb_get_summary(aoi = NULL, ...)
}
\arguments{
\item{aoi}{Area of Interest (State, District, or Bank ID)}

\item{...}{Additional arguments passed to \code{ribits()}}
}
\value{
A tibble with one row per bank
}
\description{
Convenience function to retrieve the "Master Summary" dataframe (one row per bank).
Includes all harmonized attributes, flattened spatial data, and wide-form credit metrics.
}
