% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banks.R
\name{rb_list_banks}
\alias{rb_list_banks}
\alias{get_all_banks}
\title{List mitigation banks}
\usage{
rb_list_banks(state = NULL, district = NULL)

get_all_banks()
}
\arguments{
\item{state}{Character. Two-letter state abbreviation (e.g., "FL").}

\item{district}{Character. USACE district code.}
}
\value{
A tibble containing bank information.
}
\description{
Retrieves a list of mitigation banks from RIBITS, optionally filtered by
various parameters.
}
