% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribits-user.R
\name{rb_params}
\alias{rb_params}
\title{List Available Data Parameters}
\usage{
rb_params(type = c("banks", "transactions", "contacts"))
}
\arguments{
\item{type}{Data type to list parameters for ("banks", "transactions", "contacts")}
}
\value{
A tibble with column names and descriptions
}
\description{
Lists the available data columns and metrics that can be retrieved, similar to
\code{StreamCatTools::sc_get_params()}. This helps users understand what data is available
in the "Master Summary" and "Master Ledger" tables.
}
\examples{
rb_params("banks")
rb_params("transactions")
}
