% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-core.R
\name{rb_parse_response}
\alias{rb_parse_response}
\title{Parse RIBITS API response}
\usage{
rb_parse_response(resp)
}
\arguments{
\item{resp}{A httr2 response object}
}
\value{
A tibble or list depending on the endpoint
}
\description{
Parse RIBITS API response
}
\keyword{internal}
