% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-core.R
\name{rb_process_single_item}
\alias{rb_process_single_item}
\title{Process single item response}
\usage{
rb_process_single_item(data)
}
\arguments{
\item{data}{Parsed JSON response}
}
\value{
A tibble
}
\description{
Converts API response for single item into a tibble with list-columns.
}
\keyword{internal}
