% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{rb_read_bank_summary}
\alias{rb_read_bank_summary}
\title{Read Bank Summary CSV}
\usage{
rb_read_bank_summary(path)
}
\arguments{
\item{path}{Path to the CSV file.}
}
\value{
A tibble.
}
\description{
Reads and cleans the "Bank Summary" CSV file manually downloaded from RIBITS.
}
