% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{rb_read_credit_classification}
\alias{rb_read_credit_classification}
\title{Read Credit Classification CSV}
\usage{
rb_read_credit_classification(path)
}
\arguments{
\item{path}{Path to the CSV file.}
}
\value{
A tibble.
}
\description{
Reads and cleans the "Credit Classification by Jurisdiction" CSV file.
}
