% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{rb_read_credit_tracking}
\alias{rb_read_credit_tracking}
\title{Read Credit Tracking CSV}
\usage{
rb_read_credit_tracking(path)
}
\arguments{
\item{path}{Path to the CSV file.}
}
\value{
A tibble.
}
\description{
Reads and cleans the "Bank and ILF Program Credit Tracking" CSV file.
}
