% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{rb_read_credit_withdrawal}
\alias{rb_read_credit_withdrawal}
\title{Read Credit Withdrawal CSV}
\usage{
rb_read_credit_withdrawal(path)
}
\arguments{
\item{path}{Path to the CSV file.}
}
\value{
A tibble.
}
\description{
Reads and cleans the "Bank and ILF Program Credit Withdrawal" CSV file.
}
