% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{rb_read_service_area_comments}
\alias{rb_read_service_area_comments}
\title{Read Service Area Comments CSV}
\usage{
rb_read_service_area_comments(path)
}
\arguments{
\item{path}{Path to the CSV file.}
}
\value{
A tibble.
}
\description{
Reads and cleans the "Service Area Comments" CSV file.
}
