% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{rb_report_types}
\alias{rb_report_types}
\title{List available report types}
\usage{
rb_report_types()
}
\value{
A data frame of available report types and descriptions.
}
\description{
List available report types
}
