% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-core.R
\name{rb_request}
\alias{rb_request}
\title{Make a request to the RIBITS API}
\usage{
rb_request(endpoint, params = list())
}
\arguments{
\item{endpoint}{The API endpoint (e.g., "bank_site_list/")}

\item{params}{A named list of query parameters}
}
\value{
A httr2 response object
}
\description{
Make a request to the RIBITS API
}
\keyword{internal}
