% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{REQUIRED_BANK_COLUMNS}
\alias{REQUIRED_BANK_COLUMNS}
\title{Required column names for bank data}
\format{
An object of class \code{character} of length 3.
}
\usage{
REQUIRED_BANK_COLUMNS
}
\description{
Core columns that must be present in bank dataframes.
}
\keyword{internal}
