% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{REQUIRED_GEOMETRY_COLUMNS}
\alias{REQUIRED_GEOMETRY_COLUMNS}
\title{Required column names for geometry data}
\format{
An object of class \code{character} of length 1.
}
\usage{
REQUIRED_GEOMETRY_COLUMNS
}
\description{
Core columns that must be present in spatial dataframes.
}
\keyword{internal}
