% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{REQUIRED_TRANSACTION_COLUMNS}
\alias{REQUIRED_TRANSACTION_COLUMNS}
\title{Required column names for transaction data}
\format{
An object of class \code{character} of length 2.
}
\usage{
REQUIRED_TRANSACTION_COLUMNS
}
\description{
Core columns that must be present in transaction dataframes.
}
\keyword{internal}
