% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribits-methods.R
\name{resolutions}
\alias{resolutions}
\alias{resolutions.ribits_data}
\title{Access auto-harmonization resolutions from ribits_data}
\usage{
resolutions(x, ...)

\method{resolutions}{ribits_data}(x, ...)
}
\arguments{
\item{x}{A ribits_data object}

\item{...}{Additional arguments passed to methods}
}
\value{
A tibble of harmonization resolutions
}
\description{
Access auto-harmonization resolutions from ribits_data
}
