% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\docType{data}
\name{RIBITS_REPORTS}
\alias{RIBITS_REPORTS}
\title{RIBITS Report Registry}
\format{
An object of class \code{list} of length 12.
}
\usage{
RIBITS_REPORTS
}
\description{
Mapping of friendly report names to their Oracle APEX Page IDs.
}
\keyword{internal}
