% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SQ_METERS_PER_ACRE}
\alias{SQ_METERS_PER_ACRE}
\title{Square meters per acre conversion factor}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
SQ_METERS_PER_ACRE
}
\description{
Used for converting geometry areas between units.
1 acre = 4046.86 square meters
}
\keyword{internal}
