% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribits-methods.R
\name{subset.ribits_data}
\alias{subset.ribits_data}
\title{Subset ribits_data by bank IDs}
\usage{
\method{subset}{ribits_data}(x, bank_ids, ...)
}
\arguments{
\item{x}{A ribits_data object}

\item{bank_ids}{Bank IDs to keep}

\item{...}{Additional arguments ignored}
}
\value{
Filtered ribits_data object
}
\description{
Subset ribits_data by bank IDs
}
