% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SUPPORTED_SPATIAL_FORMATS}
\alias{SUPPORTED_SPATIAL_FORMATS}
\title{Supported export formats for spatial data}
\format{
An object of class \code{character} of length 4.
}
\usage{
SUPPORTED_SPATIAL_FORMATS
}
\description{
Supported export formats for spatial data
}
\keyword{internal}
