% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALID_BANK_STATUSES}
\alias{VALID_BANK_STATUSES}
\title{Valid bank statuses}
\format{
An object of class \code{character} of length 4.
}
\usage{
VALID_BANK_STATUSES
}
\description{
Valid bank statuses
}
\keyword{internal}
