% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALID_BANK_TYPES}
\alias{VALID_BANK_TYPES}
\title{Valid bank types}
\format{
An object of class \code{character} of length 4.
}
\usage{
VALID_BANK_TYPES
}
\description{
Valid bank types
}
\keyword{internal}
