% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALID_TRANSACTION_LEVELS}
\alias{VALID_TRANSACTION_LEVELS}
\title{Valid transaction types}
\format{
An object of class \code{character} of length 3.
}
\usage{
VALID_TRANSACTION_LEVELS
}
\description{
Valid transaction types
}
\keyword{internal}
