% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribits.R
\name{arrange.ribits_data}
\alias{arrange.ribits_data}
\title{Arrange ribits_data objects}
\usage{
arrange.ribits_data(.data, ...)
}
\arguments{
\item{.data}{A ribits_data object}

\item{...}{Arrange expressions passed to dplyr::arrange()}
}
\value{
An arranged ribits_data object
}
\description{
Apply dplyr::arrange() to the banks dataframe within a ribits_data object.
}
\examples{
\dontrun{
ca_banks <- ribits(state = "CA")
sorted_banks <- ca_banks |> arrange(desc(total_acres))
}
}
