% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribits.R
\name{plot.ribits_data}
\alias{plot.ribits_data}
\title{Plot ribits_data objects}
\usage{
\method{plot}{ribits_data}(x, ...)
}
\arguments{
\item{x}{A ribits_data object}

\item{...}{Additional arguments passed to the underlying plot function}
}
\value{
NULL (called for side effects)
}
\description{
Provides a quick visual overview of the data.
\itemize{
\item If spatial data is available, it maps the geometries.
\item If no spatial data but transactions exist, it plots a credit summary.
}
}
\examples{
\dontrun{
ca <- ribits(state = "CA")
plot(ca)
}
}
