% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribits.R
\name{print.ribits_data}
\alias{print.ribits_data}
\title{Print method for ribits_data}
\usage{
\method{print}{ribits_data}(x, ...)
}
\arguments{
\item{x}{A ribits_data object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for ribits_data
}
