% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{rb_clear_checkpoints}
\alias{rb_clear_checkpoints}
\title{Clear all checkpoints (Internal)}
\usage{
rb_clear_checkpoints()
}
\description{
This function is internal. Checkpoints are cleared automatically on success.
}
\keyword{internal}
