% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rb_extract_footprint}
\alias{rb_extract_footprint}
\title{Extract bank footprint}
\usage{
rb_extract_footprint(bank)
}
\arguments{
\item{bank}{A tibble returned by \code{rb_get_bank()}.}
}
\value{
An sf object containing the footprint, or NULL if not found.
}
\description{
Extracts the bank footprint geometry from a bank object and converts it to
an sf object.
}
