% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{rb_extract_irt_members}
\alias{rb_extract_irt_members}
\title{Extract IRT members from bank object}
\usage{
rb_extract_irt_members(bank)
}
\arguments{
\item{bank}{A tibble returned by \code{rb_get_bank()}}
}
\value{
A tibble containing IRT members.
}
\description{
Extract IRT members from bank object
}
