% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{rb_extract_ledger}
\alias{rb_extract_ledger}
\title{Extract ledger from bank object}
\usage{
rb_extract_ledger(bank)
}
\arguments{
\item{bank}{A tibble returned by \code{rb_get_bank()}}
}
\value{
A tibble containing the ledger, or an empty tibble if not found.
}
\description{
Extract ledger from bank object
}
