% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{rb_extract_managers}
\alias{rb_extract_managers}
\title{Extract managers from bank object}
\usage{
rb_extract_managers(bank)
}
\arguments{
\item{bank}{A tibble returned by \code{rb_get_bank()}}
}
\value{
A tibble containing managers.
}
\description{
Extract managers from bank object
}
