% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{rb_extract_other_contacts}
\alias{rb_extract_other_contacts}
\title{Extract other contacts from RIBITS data}
\usage{
rb_extract_other_contacts(data)
}
\arguments{
\item{data}{A tibble returned by \verb{rb_get_*()} functions}
}
\value{
A tibble
}
\description{
Extract other contacts from RIBITS data
}
\keyword{internal}
