% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{rb_extract_program_sites}
\alias{rb_extract_program_sites}
\title{Extract program sites from ILF program data}
\usage{
rb_extract_program_sites(data)
}
\arguments{
\item{data}{A tibble returned by \code{rb_get_ilf_program()}}
}
\value{
A tibble
}
\description{
Extract program sites from ILF program data
}
\keyword{internal}
