% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{rb_extract_service_area}
\alias{rb_extract_service_area}
\title{Extract service area from RIBITS data}
\usage{
rb_extract_service_area(data)
}
\arguments{
\item{data}{A tibble returned by \verb{rb_get_*()} functions}
}
\value{
An sf object or empty tibble
}
\description{
Extract service area from RIBITS data
}
\keyword{internal}
