% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{rb_extract_sponsors}
\alias{rb_extract_sponsors}
\title{Extract sponsors from bank object}
\usage{
rb_extract_sponsors(bank)
}
\arguments{
\item{bank}{A tibble returned by \code{rb_get_bank()}}
}
\value{
A tibble containing sponsors.
}
\description{
Extract sponsors from bank object
}
