% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{rb_extract_umbrella_sites}
\alias{rb_extract_umbrella_sites}
\title{Extract umbrella sites from umbrella data}
\usage{
rb_extract_umbrella_sites(data)
}
\arguments{
\item{data}{A tibble returned by \code{rb_get_umbrella()}}
}
\value{
A tibble
}
\description{
Extract umbrella sites from umbrella data
}
\keyword{internal}
