% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unified-api.R
\name{rb_programs_by_location}
\alias{rb_programs_by_location}
\title{Find ILF programs by location}
\usage{
rb_programs_by_location(lat, lon)
}
\arguments{
\item{lat}{Latitude (decimal degrees)}

\item{lon}{Longitude (decimal degrees)}
}
\value{
A tibble of ILF programs near the location
}
\description{
Queries the RIBITS API to find ILF programs near a given latitude/longitude.
}
\keyword{internal}
