% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribits.R
\name{select.ribits_data}
\alias{select.ribits_data}
\title{Select columns from ribits_data objects}
\usage{
select.ribits_data(.data, ...)
}
\arguments{
\item{.data}{A ribits_data object}

\item{...}{Select expressions passed to dplyr::select()}
}
\value{
A ribits_data object with selected columns
}
\description{
Apply dplyr::select() to the banks dataframe within a ribits_data object.
}
\examples{
\dontrun{
ca_banks <- ribits(state = "CA")
simplified <- ca_banks |> select(bank_id, bank_name, total_acres)
}
}
